Accounts receivable

+++ Dtos/Sales/Invoices/InvoiceDetailDto.cs
@@ -4,12 +4,12 @@ namespace API.Dtos.Sales.Invoices
 {
     public class InvoiceDetailDto
     {
-        [Range(0, int.MaxValue, ErrorMessage = "The field {0} must be greater than {1}.")]
+        [Range(0, double.MaxValue, ErrorMessage = "The field {0} must be greater than {1}.")]
         public int ProductId { get; set; }
-        [Range(0.0, Double.MaxValue, ErrorMessage = "The field {0} must be greater than {1}.")]
+        [Range(0.0, double.MaxValue, ErrorMessage = "The field {0} must be greater than {1}.")]
         public decimal Quantity { get; set; }
         public decimal Cost { get; set; }
-        [Range(0.0, Double.MaxValue, ErrorMessage = "The field {0} must be greater than {1}.")]
+        [Range(0.0, double.MaxValue, ErrorMessage = "The field {0} must be greater than {1}.")]
         public decimal Price { get; set; }
         public decimal Disccount { get; set; }
         public decimal Total { get; set; }
